Setup new generic Target class that enemy/resource can extend
