+++ src/target.js
@@ -1,11 +1,12 @@
-import { HealthBar } from "./ui/health-bar.js";
-
 export class Target {
     target;
 	progressBar;
-    scene;
     name = "";
     drops = [];
+
+    scene;
+    stats;
+    characterData;
 
     constructor(data) {
         // Add target
@@ -13,8 +14,8 @@ export class Target {
         this.target
             .setOrigin(0.5, 0)
             .setDepth(2)
-            .setScale(0.4);
-			.setInteractive();
+            .setScale(.4)
+			.setInteractive()
 			.on("pointerup", () => {
 				// Check if it is the current target
 				if (this.target.visible) {
@@ -28,6 +29,7 @@ export class Target {
         this.drops = data.drops;
         this.scene = data.scene;
         this.stats = data.scene.stats;
+        this.characterData = data.scene.characterData;
     }
 
     clickTarget() {
@@ -42,7 +44,7 @@ export class Target {
         this.stats.updateClickedTargetStat();
     }
 
-    updateProgress(damage) {
+    updateProgress(progress) {
         let isFinished = this.progressBar.updateProgress(progress);
 
         if (isFinished) {
@@ -72,109 +74,4 @@ export class Target {
         this.target.visible = false;
         this.progressBar.hide();
     }
-}
-
-class Enemy extends Target {
-	blueHitsplat;
-    redHitsplat;
-    hitsplatText = "1";
-	killGold;
-	
-	constructor(data) {
-		super(data);
-		
-		// Add hitsplats
-        this.blueHitsplat = data.scene.add
-            .image(data.x, data.y + 50, "blue-hitsplat")
-            .setOrigin(0.5, 0)
-            .setDepth(3);
-			.setScale(0.3);
-        this.blueHitsplat.visible = false;
-
-        this.redHitsplat = data.scene.add
-            .image(data.x, data.y + 50, "red-hitsplat")
-            .setOrigin(0.5, 0)
-            .setDepth(3);
-			.setScale(0.3);
-        this.redHitsplat.visible = false;
-
-		// Add damage text
-        this.hitsplatText = data.scene.add
-			.text(data.x, data.y + 100, "1", {
-				fill: "white"
-			})
-			.setOrigin(0.5, 0)
-			.setDepth(4);
-        this.hitsplatText.visible = false;
-
-        // Add health bar
-        this.progressBar = new HealthBar(data.scene, data.x, data.y - 40, data.health);
-		
-		// Enemy specific vars
-		this.killGold = data.killGold;
-	}
-	
-	getClickValue() {
-		// Get damage based on level
-        let damageLevel =  this.scene.getDamageLevel();
-		let hitValue = Math.floor(Math.random() * (damageLevel + 1));
-		return hitValue;
-	}
-	
-	onClick(hitValue) {
-		// Get bonus gold for using mouseclick to encourage user interaction
-        this.stats.addGold(hitValue);
-		
-		// Update stats
-		this.stats.updateClickDamageStat(hitValue);
-		// Update skills here
-		
-		// Display hit
-        this.hitsplatText.text = hitValue;
-        hitValue == 0
-            ? (this.blueHitsplat.visible = true)
-            : (this.redHitsplat.visible = true);
-        this.hitsplatText.visible = true;
-		
-		// Hide hitsplat
-        let _this = this;
-        setTimeout(function() {
-            _this.redHitsplat.visible = false;
-            _this.blueHitsplat.visible = false;
-            _this.hitsplatText.visible = false;
-        }, 200);
-	}
-	
-	onCompletion() {
-		// Give extra gold if unit is killed
-		this.stats.addGold(this.killGold);
-		console.log(this.name + " killed, getting " + this.killGold + " extra gold");
-		
-		// Update quest and stats
-        this.scene.targetKilled(this.name);
-	}
-}
-
-class Resource extends Target {
-	skill;
-	
-	constructor(data) {
-		super(data);
-		
-		this.skill = data.skill;
-
-        // Add health bar
-        this.progressBar = new ProgressBar(data.scene, data.x, data.y - 40, data.health);
-	}
-	
-	getClickValue() {
-		return calcLeve(this.characterData.skills[skill]);
-	}
-	
-	onClick(clickValue) {
-		// Increase skill xp
-		this.characterData.skills[skill] += clickValue;
-	}
-	
-	onCompletion() {}
 }
